+++ index.html
@@ -14,3 +14,72 @@
     <script type="module" src="main.js"></script>
   </body>
 </html>
+
+<!-- <div class="container">
+  <ul id="list" class="comments">
+  ${usersHTML}
+  </ul>
+  ${
+    token
+      ? `<div class="add-form">
+  <input id="name-input"
+    type="text"
+    value=""
+    class="add-form-name"
+    placeholder="Введите ваше имя"
+  />
+  <textarea
+    id="comment-input"
+    class="add-form-text"
+    placeholder="Введите ваш комментарий"
+    rows="4"
+  ></textarea>
+  <div class="add-form-row">
+    <button id="add-button" class="add-form-button">Написать</button>
+  </div>
+  </div>`
+    
+  : `<p>Для добавления комментария, <a id="login-link" class="add-form-link" href='#'>зарегистрируйтесь</а></p>`
+  }
+  <div class="comment-loader hidden">
+          <span>Comment is being posted</span>
+        </div>
+  </div>
+</div>; -->
+  
+<!-- <div class="container">
+  <div id="list" class="comments">
+  <div class="password">
+   <h2 class="title"> Форма ${isLoginMode ? "Входа" : "Регистрации"}</h2>
+   ${
+     isLoginMode
+       ? ""
+       : `<input 
+   type="text"
+   class="login-input" id="name-input-authorization"
+   placeholder="Имя"
+  />
+   }</button>
+          
+          <input 
+            type="text"
+            class="login-input" id="login-input"
+            placeholder="Логин"
+          />
+          <input 
+            type="text"
+            class="password-input" id="password-input"
+            placeholder="Пароль"
+            rows="4"
+          ></textarea>
+          <div class="add-form-row">
+            <button id="enter-button" class="enter-button " >${
+              isLoginMode ? "Войти" : "Зарегистрироваться"
+            }</button>
+            <button id="signUp-button" class="enter-button " >Перейти ${
+              isLoginMode ? "К регистрации" : "Ко входу"
+            }</button></button>
+          </div> 
+  </div>
+</div>
+</div> -->
