+++ styles.css
@@ -15,14 +15,16 @@ body {
 }
 
 .comments,
-.comment {
+.comment,
+.comments1 {
   margin: 0;
   padding: 0;
   list-style: none;
 }
 
 .comment,
-.add-form {
+.add-form,
+.comments1 {
   width: 596px;
   box-sizing: border-box;
   background: radial-gradient(
@@ -34,13 +36,15 @@ body {
   border-radius: 20px;
 }
 
-.comments {
+.comments,
+.comments1 {
   display: flex;
   flex-direction: column;
   gap: 24px;
 }
 
-.comment {
+.comment,
+.comments1 {
   padding: 48px;
 }
 
@@ -122,6 +126,11 @@ body {
   justify-content: flex-end;
 }
 
+.add-form-row {
+  display: flex;
+  justify-content: space-around;
+}
+
 .add-form-button {
   margin-top: 24px;
   font-size: 24px;
@@ -154,4 +163,13 @@ body {
 .comment__link {
   text-decoration: none;
   color: white;
+}
+.enter-button {
+  margin-top: 24px;
+  font-size: 24px;
+  padding: 10px 20px;
+  background-color: #bcec30;
+  border: none;
+  border-radius: 18px;
+  cursor: pointer;
 }
